requirements.txt
awkward
autopep8
black
boost-histogram
docopt
hydra-core
jupyter
jupyter-book
lightning
matplotlib
mplhep
notebook
numpy
papermill
pre-commit
scikit-learn
scikit-optimize
scipy
seaborn
setGPU
tensorboard
torch
torch-geometric
tqdm
uproot
vector
wheel
xgboost